requirements.txt
anthropic>=0.66.0
pinecone>=5.4.0
tiktoken>=0.5.0
openai>=1.30.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyGithub>=1.59.0
aiofiles>=23.0.0
trafilatura>=1.6.0
